requirements.txt
click==8.1.7
fabric==3.2.2
GitPython==3.1.36
watchdog==3.0.0
